Refactor gRPC out of Webjobs.Script
