+++ src/WebJobs.Script.Grpc/Eventing/RpcEvent.cs
@@ -1,9 +1,10 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+using Microsoft.Azure.WebJobs.Script.Eventing;
 using Microsoft.Azure.WebJobs.Script.Grpc.Messages;
 
-namespace Microsoft.Azure.WebJobs.Script.Eventing.Rpc
+namespace Microsoft.Azure.WebJobs.Script.Grpc.Eventing
 {
     public class RpcEvent : ScriptEvent
     {
