+++ src/WebJobs.Script.Grpc/MessageExtensions/RpcException.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.Azure.WebJobs.Script.Workers.Rpc
+namespace Microsoft.Azure.WebJobs.Script.Grpc
 {
     public class RpcException : Exception
     {
