+++ src/WebJobs.Script.Grpc/MessageExtensions/RpcMessageExtensionUtilities.cs
@@ -8,7 +8,7 @@ using System.Linq;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Azure.WebJobs.Script.Grpc.Messages;
 
-namespace Microsoft.Azure.WebJobs.Script.Workers.Rpc
+namespace Microsoft.Azure.WebJobs.Script.Grpc
 {
     internal static class RpcMessageExtensionUtilities
     {
