+++ src/WebJobs.Script.Grpc/Server/IRpcServer.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Microsoft.Azure.WebJobs.Script.Abstractions
+namespace Microsoft.Azure.WebJobs.Script.Workers.Rpc
 {
     public interface IRpcServer
     {
