+++ src/WebJobs.Script/Workers/ProcessManagement/IProcessRegistry.cs
@@ -6,7 +6,7 @@ using System.Diagnostics;
 
 namespace Microsoft.Azure.WebJobs.Script.Workers
 {
-    internal interface IProcessRegistry
+    public interface IProcessRegistry
     {
         // Registers processes to ensure that they are cleaned up on host exit.
         bool Register(Process process);
